+++ flutter_tide/example/lib/services.dart
@@ -1,0 +1,12 @@
+abstract class ServerClient {
+  const ServerClient();
+  Future<int> getValue();
+}
+
+class MockServerClient extends ServerClient {
+  const MockServerClient();
+  Future<int> getValue() async {
+    await Future.delayed(const Duration(seconds: 2));
+    return 128;
+  }
+}