+++ tide/example/tide_example.dart
@@ -22,7 +22,10 @@ class CounterState {
 class Increment extends StoreAction<CounterState> {
   @override
   Stream<CounterState> execute(
-      StateReader<CounterState> state, Dispatcher dispatch) async* {
+    StateReader<CounterState> state,
+    Dispatcher dispatch,
+    ServiceLocator services,
+  ) async* {
     if (!state().isLoading) {
       yield CounterState(state().value + 1, false);
     }
@@ -32,7 +35,10 @@ class Increment extends StoreAction<CounterState> {
 class AddServerValue extends StoreAction<CounterState> {
   @override
   Stream<CounterState> execute(
-      StateReader<CounterState> state, Dispatcher dispatch) async* {
+    StateReader<CounterState> state,
+    Dispatcher dispatch,
+    ServiceLocator services,
+  ) async* {
     if (!state().isLoading) {
       yield CounterState(state().value, true);
       final serverValue = await const ServerClient().getValue();
@@ -47,7 +53,10 @@ class ResetThenAddValueverySecond extends StoreAction<CounterState> {
 
   @override
   Stream<CounterState> execute(
-      StateReader<CounterState> state, Dispatcher dispatch) async* {
+    StateReader<CounterState> state,
+    Dispatcher dispatch,
+    ServiceLocator services,
+  ) async* {
     if (!state().isLoading) {
       yield CounterState(0, true);
       for (var i = 0; i < 5; i++) {