+++ tide/lib/tide.dart
@@ -4,3 +4,4 @@ export 'package:meta/meta.dart';
 
 export 'src/action.dart';
 export 'src/store.dart';
+export 'src/service.dart';